+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Add support for publishing Java and Kotlin Protobuf compiled sources to Maven Central [#562](https://github.com/provenance-io/provenance/issues/562)
 * Adds support for creating root name governance proposals from the cli [#599](https://github.com/provenance-io/provenance/issues/599)
 * Adding of the msg based fee module [#354](https://github.com/provenance-io/provenance/issues/354)
+* Upgrade provenance to 0.45 cosmos sdk release [#607](https://github.com/provenance-io/provenance/issues/607)
 
 ### Improvements